+++ NEWS.md
@@ -8,6 +8,7 @@
 * `sub_string()`: Can extract text from left, right or middle and is able to look up characters as start or end position. (22.01.2026)
 * `remove_blanks()`: Removes blanks in the expressions of a character variable. (24.01.2026)
 * `macro()`, `apply_macro()`: Resolve objects starting with "&" within a text. Functions which can print titles, footnotes, variable or statistic labels make use of these functions, so that e.g. a title can be passed as "The current year is &year" and will be resolved to "The current year is 2026". (24.01.2026)
+* `where.()`: A quick way to filter a data frame and get a direct view of the result. (27.01.2026)
 
 ### New functionality
 
@@ -51,6 +52,7 @@
 ### Additionally
 
 * `any_table()`: Removed c() in examples where not necessary. (24.01.2026)
+* `if.()`: Now outputs a message on how many observations have been removed and how many are left. (27.01.2026)
 
 
 # qol 1.2.0